Added start code for Humphrey
This code is intended to be used at the start of every g-code. Can be
incorpareted into post processors in various CAM programs. It will
intiate go to absolute home postion of the oriental motors, probe the
bit length with the Z axis, then adjust for material thickness and bed
offset before it starts the spindle