fixed motor plate adjustment holes
now they allow the pinion to be moved completely clear of the rack. this
allows the for easier tuning of the glide bock positions. if the pinoin
engages the rack, the resitance from the motor makes it more difficult
to feel the amount of friction and play from the glideblock adjustment